Module effektio_core::events

Structs

Color

The color

Colorize

Customize the color scheme

CommentEventDevContent

The payload for our Comment event.

EventId

A Matrix event ID.

ImageMessageEventContent

The payload for an image message.

Labels

NewsEventDevContent

The payload for our news event.

TaskDevContent

The Task Event

TaskListDevContent

The TaskList Event

TextMessageEventContent

The payload for a text message.

VideoMessageEventContent

The payload for a video message.

Enums

CommentEvent

The content that is specific to each Comment type variant.

NewsContentType

The content that is specific to each message type variant.

NewsEvent

The content that is specific to each news type variant.

SpecialTaskListRole

Task

The content that is specific to each news type variant.

TaskList

The content that is specific to each news type variant.

TaskPriority

Implementing Priority according to https://www.rfc-editor.org/rfc/rfc8984.html#section-4.4.1

TimeZone

TimeZones built at compile time from the tz database

Type Definitions

BelongsTo

UtcDateTime

Default UTC Datetime Object